V4 merged express

+++ ai_workspace/api_views.py
@@ -1666,12 +1666,14 @@ class TaskAssignInfoCreateView(viewsets.ViewSet):
                     print("Inside Else")
                     reassigns = TaskAssign.objects.filter(Q(task=obj.task_assign.task) & Q(step=obj.task_assign.step) & Q(reassigned = True))
                     if reassigns:
-                        obj_1 = reassigns.first().task_assign_info
-                        self.history(obj_1)
-                        obj_1.task_assign.assign_to = user
-                        obj_1.task_assign.status = 1
-                        obj_1.task_assign.save()
-                        obj_1.delete()
+                        try:obj_1 = reassigns.first().task_assign_info
+                        except:obj_1=None
+                        if obj_1:
+                            self.history(obj_1)
+                            obj_1.task_assign.assign_to = user
+                            obj_1.task_assign.status = 1
+                            obj_1.task_assign.save()
+                            obj_1.delete()
                     #assigned_user = obj.task_assign.assign_to
                     obj.task_assign.assign_to = user
                     obj.task_assign.status = 1
